requirements.txt
setuptools
wheel
taipy
google-generativeai
python-dotenv